fix package name & delete dependency profiler & fix tab to space
