+++ plugins/json-lib/src/main/java/com/navercorp/pinpoint/plugin/json/lib/JsonLibPlugin.java
@@ -41,11 +41,9 @@ public class JsonLibPlugin implements ProfilerPlugin {
     
     private void addJSONSerializerInterceptor(ProfilerPluginContext context) {
         ClassFileTransformerBuilder builder = context.getClassFileTransformerBuilder("net.sf.json.JSONSerializer");
+        builder.editMethod("toJSON", "java.lang.Object").injectInterceptor("com.navercorp.pinpoint.bootstrap.interceptor.BasicMethodInterceptor", JsonLibConstants.SERVICE_TYPE);
         
-	builder.editMethod("toJSON", "java.lang.Object").injectInterceptor("com.navercorp.pinpoint.bootstrap.interceptor.BasicMethodInterceptor", JsonLibConstants.SERVICE_TYPE);
-        
-	builder.editMethod("toJava", "net.sf.json.JSON").injectInterceptor("com.navercorp.pinpoint.bootstrap.interceptor.BasicMethodInterceptor", JsonLibConstants.SERVICE_TYPE);
-        
+        builder.editMethod("toJava", "net.sf.json.JSON").injectInterceptor("com.navercorp.pinpoint.bootstrap.interceptor.BasicMethodInterceptor", JsonLibConstants.SERVICE_TYPE);
         ClassFileTransformer transformer = builder.build();
         context.addClassFileTransformer(transformer);
     }
@@ -53,9 +51,9 @@ public class JsonLibPlugin implements ProfilerPlugin {
     private void addJSONObjectInterceptor(ProfilerPluginContext context) {
         ClassFileTransformerBuilder builder = context.getClassFileTransformerBuilder("net.sf.json.JSONObject");
         
-	builder.editMethod("fromObject", "java.lang.Object").injectInterceptor("com.navercorp.pinpoint.bootstrap.interceptor.BasicMethodInterceptor", JsonLibConstants.SERVICE_TYPE);
+        builder.editMethod("fromObject", "java.lang.Object").injectInterceptor("com.navercorp.pinpoint.bootstrap.interceptor.BasicMethodInterceptor", JsonLibConstants.SERVICE_TYPE);
         
-	builder.editMethod("toBean", "net.sf.json.JSONObject").injectInterceptor("com.navercorp.pinpoint.bootstrap.interceptor.BasicMethodInterceptor", JsonLibConstants.SERVICE_TYPE);
+        builder.editMethod("toBean", "net.sf.json.JSONObject").injectInterceptor("com.navercorp.pinpoint.bootstrap.interceptor.BasicMethodInterceptor", JsonLibConstants.SERVICE_TYPE);
         
         ClassFileTransformer transformer = builder.build();
         context.addClassFileTransformer(transformer);
@@ -65,7 +63,7 @@ public class JsonLibPlugin implements ProfilerPlugin {
         ClassFileTransformerBuilder builder = context.getClassFileTransformerBuilder("net.sf.json.JSONArray");
         builder.editMethod("fromObject", "java.lang.Object").injectInterceptor("com.navercorp.pinpoint.bootstrap.interceptor.BasicMethodInterceptor", JsonLibConstants.SERVICE_TYPE);
         
-	builder.editMethod("toArray", "net.sf.json.JSONArray").injectInterceptor("com.navercorp.pinpoint.bootstrap.interceptor.BasicMethodInterceptor", JsonLibConstants.SERVICE_TYPE);
+        builder.editMethod("toArray", "net.sf.json.JSONArray").injectInterceptor("com.navercorp.pinpoint.bootstrap.interceptor.BasicMethodInterceptor", JsonLibConstants.SERVICE_TYPE);
         
         ClassFileTransformer transformer = builder.build();
         context.addClassFileTransformer(transformer);
